Address outstanding PR #81 review comments
