+++ src/App/SoloDevBoard.App/Components/Layout/MainLayout.razor
@@ -32,7 +32,7 @@
 @code {
     private bool _drawerOpen = true;
     private bool _isDarkMode = false;
-    private readonly MudTheme _theme = SoloDevBoardTheme.MudTheme;
+    private readonly MudTheme _theme = MainLayoutTheme.MudTheme;
 
     private void DrawerToggle()
     {
